Move test Fixture to separate class for easier reuse
